CAMEL-21295: refactor knative kubernetes export test
